Langchain_tool_calling_agent/requirements.txt
langchain
wikipedia
python-dotenv
langchain-core
langchain-community
pydantic
langchain-openai
langchain_huggingface
duckduckgo-search
wikipedia